#1698 use decimal for components

+++ src/Frontend/FluentCMS.Web.UI.Components/Components/Form/FormCheckboxGroup/FormCheckboxGroup.razor.cs
@@ -5,7 +5,7 @@ namespace FluentCMS.Web.UI.Components;
 public partial class FormCheckboxGroup<TItem, TValue>
 {
     [Parameter]
-    public int Cols { get; set; } = 12;
+    public decimal Cols { get; set; } = 12;
 
     [Parameter]
     public ICollection<TItem> Data { get; set; } = [];
